break getFieldFromValue into smaller functions based on the types of fields being pushed
